refactor: isolate developer-tools module in data layer
Move all developer tools fetching logic into a dedicated folder to
maintain data layer isolation and prevent bidirectional imports
between data layer and app code.

Changes:
- Create src/data-layer/fetchers/developer-tools/ module with:
  - index.ts: main fetchDeveloperTools orchestrator
  - utils.ts: types, constants, and utility functions
  - fetchBuidlGuidl.ts: BuidlGuidl API fetcher
  - fetchGitHub.ts: GitHub GraphQL enrichment
  - fetchNpmJs.ts: npm downloads enrichment

- Move envelope types (DeveloperToolsDataEnvelope,
  DeveloperAppsComputedSelections) from app code to data layer

- Keep only UI-specific code in app (getCategoryTagStyle)

- Update all imports to use the new module structure

This ensures the data layer remains self-contained and doesn't depend
on app-level code for runtime functions.
