+++ src/data-layer/index.ts
@@ -15,8 +15,7 @@ import type {
 } from "@/lib/types"
 import type { CommunityEventsReturnType } from "@/lib/interfaces"
 
-import type { DeveloperToolsDataEnvelope } from "../../app/[locale]/developers/apps/types"
-
+import type { DeveloperToolsDataEnvelope } from "./fetchers/developer-tools/utils"
 import type { BeaconChainData } from "./fetchers/fetchBeaconChain"
 import type { CoinGeckoCoinMarketResponse } from "./fetchers/fetchStablecoinsData"
 import { get } from "./storage"
@@ -45,5 +44,4 @@ export const getStablecoinsData = () => get<CoinGeckoCoinMarketResponse>(KEYS.ST
 export const getTotalEthStakedData = () => get<MetricReturnData>(KEYS.TOTAL_ETH_STAKED)
 export const getTotalValueLockedData = () => get<MetricReturnData>(KEYS.TOTAL_VALUE_LOCKED)
 export const getEventsData = () => get<EventItem[]>(KEYS.EVENTS)
-export const getDeveloperToolsData = () =>
-  get<DeveloperToolsDataEnvelope>(KEYS.DEVELOPER_APPS)
+export const getDeveloperToolsData = () => get<DeveloperToolsDataEnvelope>(KEYS.DEVELOPER_APPS)
